Rename submodules to remove unnecessary 'edfi_' prefixes.
