+++ edfi_api_client/__init__.py
@@ -1,2 +1,2 @@
-from edfi_api_client.edfi_client import EdFiClient
+from edfi_api_client.client import EdFiClient
 from edfi_api_client.util import camel_to_snake, snake_to_camel, url_join
