+++ edfi_api_client/async_mixin.py
@@ -6,13 +6,13 @@ import logging
 from requests.exceptions import RequestsWarning
 
 from edfi_api_client import util
-from edfi_api_client.edfi_session import EdFiSession
+from edfi_api_client.session import EdFiSession
 
 from typing import AsyncIterator, List, Optional
 from typing import TYPE_CHECKING
 if TYPE_CHECKING:
     from edfi_api_client import EdFiClient
-    from edfi_api_client.edfi_params import EdFiParams
+    from edfi_api_client.params import EdFiParams
 
 
 class AsyncEdFiSession(EdFiSession):
